First commit of the new Repository implementation
In working towards the integration of an object store
as a potential replacement of the current simple filesystem
implementation for the repository, new classes are defined
NodeRepository, Repository, that will implement the required
functionality.

The Repository class will actually implement the required
methods to store and retrieve files using either the file
system or an object store like Swift. The NodeRepository
will provide a layer the Node objects can use to interact
with the Repository.