Introduction of new Database and ORM models
New tables are introduced to the database model:

  * DbRepository
  * DbFile
  * DbNodeFile

This new implementation requires a node to instantiate
the NodeRepository class. This object will use an instance
of Repotable to get, put and delete files and directories
in the database. The actual content of files will be stored
physically in the repository through a Repository object.

Eventually the applicable Repository should be dynamically
created through a factory based on the configuration of the
environment. Currently is partly hard coded.