Change the way the repository is configured
Since in the near future we will support the use of multiple
repositories with potentially different implementations, the
configuration of the repository for the available profiles in
config.json needed to be adapted. Instead of defining a fixed
number of variables directly in each profile, now a single
variable REPOSITORY_NAME will be defined. This is the key that
will correspond to the desired repository, which will be
separately configured in under the 'repositories' key.
This name will also correspond to the value recorded in the
database. In this way, we should be able to check whether
the name potentially conflicts with a different repository.

Currently the parsing and validation of the configured repository
happens in aiida/settings.py, which should at some point
probably be moved to a more specific location and subsequently
imported where needed.

This new functionality assumes that any configured repository
is properly initialized and the required resources, e.g. the
directory and uuid file being readable in the case of the filesystem,
are accessible. The code that will actually be able to configure
and initialize new repositories is still to be implemented