Refactor the configuration of Repository implementation
The old version hard coded the FileSystem implementation
to be used for every Repository instance, but now that the
Swift solution is also implemented, that implementation
also needs to be tested.

Additionally, a design choice is to not allow, or at the
very least discourage, multiple profiles to use one and
the same profile. Therefore in the new config.json structure
each profile will have to fully define the repositories it
should have at its disposal.